Add files via upload
This is more of a “guess the word” game. In the game, users have to enter letter guesses, and each user will have a limited number of guesses (a counter variable is needed for limiting the guesses). You can create a pre-organized list of words that users can grab words from.